remove SystemTest tests to ApplicationSystemTestCase
