+++ test/system/notification_settings_test.rb
@@ -1,7 +1,7 @@
-require "test_helper"
+require "application_system_test_case"
 require "capybara/minitest"
 
-class NotificationSettingsTest < SystemTest
+class NotificationSettingsTest < ApplicationSystemTestCase
   include Capybara::Minitest::Assertions
 
   test "changing email notification settings" do
