+++ test/system/owner_test.rb
@@ -1,6 +1,6 @@
-require "test_helper"
+require "application_system_test_case"
 
-class OwnerTest < SystemTest
+class OwnerTest < ApplicationSystemTestCase
   include ActiveJob::TestHelper
   include RubygemsHelper
