+++ test/system/profile_test.rb
@@ -1,5 +1,4 @@
 require "application_system_test_case"
-require "test_helper"
 
 class ProfileTest < ApplicationSystemTestCase
   include ActiveJob::TestHelper
